pyproject.toml
[tool.poetry]
name = "HnMChallenge"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
kaggle = "^1.5.12"
pandas = "^1.4.0"
scikit-learn = "^1.0.2"
python-dotenv = "^0.19.2"
jupyter = "^1.0.0"
matplotlib = "^3.5.1"
seaborn = "^0.11.2"
pyarrow = "^7.0.0"
sparsesvd = "^0.2.2"
pathos = "^0.2.8"
p-tqdm = "^1.3.3"
torch = "^1.10.2"
xgboost = "^1.5.2"
similaripy = "^0.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
